retry requests that fail to parse
The Campfire API often returns completely invalid HTML responses with an HTTP 200 code, which means the errback is not triggered. if an empty hash is passed through to (say) a request for a user, the user data will be empty. The errors are transient, so if we try again, we should get the JSON we're looking for. If not, we'll disconnect at max_retries.